Trying to add mouse input

+++ engine/physics/kernels.h
@@ -8,7 +8,7 @@ namespace Physics
 		{
 			if (dist < radius)
 			{
-				float volume = 6 / (glm::pi<float>() * pow(radius, 4));
+				float volume = 6 / (glm::pi<float>() * powf(radius, 4));
 				float v = radius - dist;
 				return v * v * volume;
 			}
@@ -19,7 +19,7 @@ namespace Physics
 		{
 			if (dist < radius)
 			{
-				float volume = 10.0f / (glm::pi<float>() * pow(radius, 5));
+				float volume = 10.0f / (glm::pi<float>() * powf(radius, 5));
 				float v = radius - dist;
 				return v * v * v * volume;
 			}
@@ -31,7 +31,7 @@ namespace Physics
 		{
 			if (dist <= radius)
 			{
-				float scale = 12 / (pow(radius, 4) * glm::pi<float>());
+				float scale = 12.0f / (powf(radius, 4) * glm::pi<float>());
 				float v = (radius - dist);
 				return -v * scale;
 			}
@@ -42,7 +42,7 @@ namespace Physics
 		{
 			if (dist <= radius)
 			{
-				float scale = 30 / (pow(radius, 5) * glm::pi<float>());
+				float scale = 30 / (powf(radius, 5) * glm::pi<float>());
 				float v = (radius - dist);
 				return -v * v * scale;
 			}
@@ -53,7 +53,7 @@ namespace Physics
 		{
 			if (dist < radius)
 			{
-				float scale = 4.0f / (glm::pi<float>() * pow(radius, 8));
+				float scale = 4.0f / (glm::pi<float>() * powf(radius, 8));
 				float v = radius * radius - dist * dist;
 				return v * v * v * scale;
 			}
